+++ src/application/hart1/u54_1.c
@@ -11,6 +11,7 @@
 
 #include "mpfs_hal/mss_hal.h"
 #include "inc/demo_main.h"
+#include "inc/rtt_demo.h"
 
 volatile uint32_t count_sw_ints_h1 = 0U;
 
@@ -48,7 +49,8 @@ void u54_1(void)
 
     if(MPFS_HAL_FIRST_HART == hartid)
     {
-        start_demo();
+        // start_demo();
+        start_rtt_demo();
         /* Never return */
     }