+++ src/application/inc/rtt_demo.h
@@ -1,0 +1,6 @@
+﻿#ifndef RTT_DEMO_H_
+#define RTT_DEMO_H_
+
+void start_rtt_demo(void);
+
+#endif