+++ src/middleware/rpmsg/rpmsg_lite/porting/platform/microchip/miv/rpmsg_platform.c
@@ -20,7 +20,7 @@
 #include "timers.h"
 #include "semphr.h"
 #else
-#include "utils.h"
+// #include "utils.h"
 #endif
 
 #ifdef REMOTEPROC